refactor: extract menu frontend options to minarch_frontend_opts.c/h
Moves MenuList/MenuItem type system, Menu_message/messageWithFont utilities,
and all Option*_openMenu handlers from minarch.c into a new module.
